Combat

+++ WrenClient/Source/UI/UICharacterHUD.cpp
@@ -66,7 +66,7 @@ void UICharacterHUD::Draw()
 
 	d2dDeviceContext->DrawTextLayout(D2D1::Point2F(position.x + 86.0f, position.y + 16.0f), nameTextLayout.Get(), nameBrush);
 
-	const auto healthPercent = statsComponent.health / statsComponent.maxHealth;
+	const auto healthPercent = (float)statsComponent.health / (float)statsComponent.maxHealth;
 	auto statPosX = 232.0f * healthPercent;
 	d2dFactory->CreateRectangleGeometry(D2D1::RectF(position.x + 88.0f, position.y + 34.0f, position.x + statPosX, position.y + 44.0f), healthGeometry.ReleaseAndGetAddressOf());
 
@@ -74,7 +74,7 @@ void UICharacterHUD::Draw()
 	d2dDeviceContext->FillGeometry(healthGeometry.Get(), healthBrush);
 	d2dDeviceContext->DrawGeometry(maxHealthGeometry.Get(), nameBrush, 2.0f);
 
-	const auto manaPercent = statsComponent.mana / statsComponent.maxMana;
+	const auto manaPercent = (float)statsComponent.mana / (float)statsComponent.maxMana;
 	statPosX = 232.0f * manaPercent;
 	d2dFactory->CreateRectangleGeometry(D2D1::RectF(position.x + 88.0f, position.y + 48.0f, position.x + statPosX, position.y + 58.0f), manaGeometry.ReleaseAndGetAddressOf());
 
@@ -82,7 +82,7 @@ void UICharacterHUD::Draw()
 	d2dDeviceContext->FillGeometry(manaGeometry.Get(), manaBrush);
 	d2dDeviceContext->DrawGeometry(maxManaGeometry.Get(), nameBrush, 2.0f);
 
-	const auto staminaPercent = statsComponent.stamina / statsComponent.maxStamina;
+	const auto staminaPercent = (float)statsComponent.stamina / (float)statsComponent.maxStamina;
 	statPosX = 232.0f * staminaPercent;
 	d2dFactory->CreateRectangleGeometry(D2D1::RectF(position.x + 88.0f, position.y + 62.0f, position.x + statPosX, position.y + 72.0f), staminaGeometry.ReleaseAndGetAddressOf());
